System.currentSystemInMillis() returns long

+++ src/main/java/org/team114/ocelot/logging/Loggable.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public interface Loggable {
 
     public UUID getId();
-    public double getTimestamp();
+    public long getTimestamp();
 
 
 }
